requirements.txt
fastapi
uvicorn
pandas
python-dotenv
google-auth
google-auth-oauthlib
google-auth-httplib2
google-auth-oauthlib
google-api-python-client
